Create linalg block header definition
